+++ rancher2/schema_cluster_rke_config_nodes.go
@@ -6,6 +6,7 @@ import (
 )
 
 var (
+	// RKEConfigNodesRoles available RKE roles for nodes
 	RKEConfigNodesRoles = []string{"controlplane", "etcd", "worker"}
 )
