+++ rancher2/schema_cluster_rke_config_services.go
@@ -6,364 +6,6 @@ import (
 
 //Schemas
 
-func clusterRKEConfigServicesSchedulerFields() map[string]*schema.Schema {
-	s := map[string]*schema.Schema{
-		"extra_args": {
-			Type:     schema.TypeMap,
-			Optional: true,
-			Computed: true,
-		},
-		"extra_binds": {
-			Type:     schema.TypeList,
-			Optional: true,
-			Elem: &schema.Schema{
-				Type: schema.TypeString,
-			},
-		},
-		"extra_env": {
-			Type:     schema.TypeList,
-			Optional: true,
-			Elem: &schema.Schema{
-				Type: schema.TypeString,
-			},
-		},
-		"image": {
-			Type:     schema.TypeString,
-			Optional: true,
-			Computed: true,
-		},
-	}
-	return s
-}
-
-func clusterRKEConfigServicesKubeproxyFields() map[string]*schema.Schema {
-	s := map[string]*schema.Schema{
-		"extra_args": {
-			Type:     schema.TypeMap,
-			Optional: true,
-			Computed: true,
-		},
-		"extra_binds": {
-			Type:     schema.TypeList,
-			Optional: true,
-			Elem: &schema.Schema{
-				Type: schema.TypeString,
-			},
-		},
-		"extra_env": {
-			Type:     schema.TypeList,
-			Optional: true,
-			Elem: &schema.Schema{
-				Type: schema.TypeString,
-			},
-		},
-		"image": {
-			Type:     schema.TypeString,
-			Optional: true,
-			Computed: true,
-		},
-	}
-	return s
-}
-
-func clusterRKEConfigServicesKubeletFields() map[string]*schema.Schema {
-	s := map[string]*schema.Schema{
-		"cluster_dns_server": {
-			Type:     schema.TypeString,
-			Optional: true,
-			Computed: true,
-		},
-		"cluster_domain": {
-			Type:     schema.TypeString,
-			Optional: true,
-			Computed: true,
-		},
-		"extra_args": {
-			Type:     schema.TypeMap,
-			Optional: true,
-			Computed: true,
-		},
-		"extra_binds": {
-			Type:     schema.TypeList,
-			Optional: true,
-			Elem: &schema.Schema{
-				Type: schema.TypeString,
-			},
-		},
-		"extra_env": {
-			Type:     schema.TypeList,
-			Optional: true,
-			Elem: &schema.Schema{
-				Type: schema.TypeString,
-			},
-		},
-		"fail_swap_on": {
-			Type:     schema.TypeBool,
-			Optional: true,
-			Computed: true,
-		},
-		"image": {
-			Type:     schema.TypeString,
-			Optional: true,
-			Computed: true,
-		},
-		"infra_container_image": {
-			Type:     schema.TypeString,
-			Optional: true,
-			Computed: true,
-		},
-	}
-	return s
-}
-
-func clusterRKEConfigServicesKubeControllerFields() map[string]*schema.Schema {
-	s := map[string]*schema.Schema{
-		"cluster_cidr": {
-			Type:     schema.TypeString,
-			Optional: true,
-			Computed: true,
-		},
-		"extra_args": {
-			Type:     schema.TypeMap,
-			Optional: true,
-			Computed: true,
-		},
-		"extra_binds": {
-			Type:     schema.TypeList,
-			Optional: true,
-			Elem: &schema.Schema{
-				Type: schema.TypeString,
-			},
-		},
-		"extra_env": {
-			Type:     schema.TypeList,
-			Optional: true,
-			Elem: &schema.Schema{
-				Type: schema.TypeString,
-			},
-		},
-		"image": {
-			Type:     schema.TypeString,
-			Optional: true,
-			Computed: true,
-		},
-		"service_cluster_ip_range": {
-			Type:     schema.TypeString,
-			Optional: true,
-			Computed: true,
-		},
-	}
-	return s
-}
-
-func clusterRKEConfigServicesKubeAPIFields() map[string]*schema.Schema {
-	s := map[string]*schema.Schema{
-		"always_pull_images": {
-			Type:     schema.TypeBool,
-			Optional: true,
-			Default:  false,
-		},
-		"extra_args": {
-			Type:     schema.TypeMap,
-			Optional: true,
-			Computed: true,
-		},
-		"extra_binds": {
-			Type:     schema.TypeList,
-			Optional: true,
-			Elem: &schema.Schema{
-				Type: schema.TypeString,
-			},
-		},
-		"extra_env": {
-			Type:     schema.TypeList,
-			Optional: true,
-			Elem: &schema.Schema{
-				Type: schema.TypeString,
-			},
-		},
-		"image": {
-			Type:     schema.TypeString,
-			Optional: true,
-			Computed: true,
-		},
-		"pod_security_policy": {
-			Type:     schema.TypeBool,
-			Optional: true,
-			Default:  false,
-		},
-		"service_cluster_ip_range": {
-			Type:     schema.TypeString,
-			Optional: true,
-			Computed: true,
-		},
-		"service_node_port_range": {
-			Type:     schema.TypeString,
-			Optional: true,
-			Computed: true,
-		},
-	}
-	return s
-}
-
-func clusterRKEConfigServicesEtcdBackupConfigS3Fields() map[string]*schema.Schema {
-	s := map[string]*schema.Schema{
-		"access_key": {
-			Type:      schema.TypeString,
-			Optional:  true,
-			Sensitive: true,
-		},
-		"bucket_name": {
-			Type:     schema.TypeString,
-			Required: true,
-		},
-		"custom_ca": {
-			Type:     schema.TypeString,
-			Optional: true,
-		},
-		"endpoint": {
-			Type:     schema.TypeString,
-			Required: true,
-		},
-		"folder": {
-			Type:     schema.TypeString,
-			Optional: true,
-		},
-		"region": {
-			Type:     schema.TypeString,
-			Optional: true,
-		},
-		"secret_key": {
-			Type:      schema.TypeString,
-			Optional:  true,
-			Sensitive: true,
-		},
-	}
-	return s
-}
-
-func clusterRKEConfigServicesEtcdBackupConfigFields() map[string]*schema.Schema {
-	s := map[string]*schema.Schema{
-		"enabled": {
-			Type:     schema.TypeBool,
-			Optional: true,
-			Default:  true,
-		},
-		"interval_hours": {
-			Type:     schema.TypeInt,
-			Optional: true,
-			Default:  12,
-		},
-		"retention": {
-			Type:     schema.TypeInt,
-			Optional: true,
-			Default:  6,
-		},
-		"s3_backup_config": {
-			Type:     schema.TypeList,
-			MaxItems: 1,
-			Optional: true,
-			Elem: &schema.Resource{
-				Schema: clusterRKEConfigServicesEtcdBackupConfigS3Fields(),
-			},
-		},
-	}
-	return s
-}
-
-func clusterRKEConfigServicesEtcdFields() map[string]*schema.Schema {
-	s := map[string]*schema.Schema{
-		"backup_config": {
-			Type:     schema.TypeList,
-			MaxItems: 1,
-			Optional: true,
-			Computed: true,
-			Elem: &schema.Resource{
-				Schema: clusterRKEConfigServicesEtcdBackupConfigFields(),
-			},
-		},
-		"ca_cert": {
-			Type:     schema.TypeString,
-			Optional: true,
-			Computed: true,
-		},
-		"cert": {
-			Type:      schema.TypeString,
-			Optional:  true,
-			Computed:  true,
-			Sensitive: true,
-		},
-		"creation": {
-			Type:     schema.TypeString,
-			Optional: true,
-			Computed: true,
-		},
-		"external_urls": {
-			Type:     schema.TypeList,
-			Optional: true,
-			Elem: &schema.Schema{
-				Type: schema.TypeString,
-			},
-		},
-		"extra_args": {
-			Type:     schema.TypeMap,
-			Optional: true,
-			Computed: true,
-		},
-		"extra_binds": {
-			Type:     schema.TypeList,
-			Optional: true,
-			Elem: &schema.Schema{
-				Type: schema.TypeString,
-			},
-		},
-		"extra_env": {
-			Type:     schema.TypeList,
-			Optional: true,
-			Elem: &schema.Schema{
-				Type: schema.TypeString,
-			},
-		},
-		"gid": {
-			Type:     schema.TypeInt,
-			Optional: true,
-			Default:  0,
-		},
-		"image": {
-			Type:     schema.TypeString,
-			Optional: true,
-			Computed: true,
-		},
-		"key": {
-			Type:      schema.TypeString,
-			Optional:  true,
-			Computed:  true,
-			Sensitive: true,
-		},
-		"path": {
-			Type:     schema.TypeString,
-			Optional: true,
-			Computed: true,
-		},
-		"retention": {
-			Type:     schema.TypeString,
-			Optional: true,
-			Computed: true,
-		},
-		"snapshot": {
-			Type:     schema.TypeBool,
-			Optional: true,
-			Computed: true,
-		},
-		"uid": {
-			Type:     schema.TypeInt,
-			Optional: true,
-			Default:  0,
-		},
-	}
-	return s
-}
-
 func clusterRKEConfigServicesFields() map[string]*schema.Schema {
 	s := map[string]*schema.Schema{
 		"etcd": {
