+++ rancher2/schema_cluster_rke_config_services_kube_api.go
@@ -1,0 +1,169 @@
+package rancher2
+
+import (
+	"github.com/hashicorp/terraform-plugin-sdk/helper/schema"
+)
+
+//Schemas
+
+func clusterRKEConfigServicesKubeAPIAuditLogConfigFields() map[string]*schema.Schema {
+	s := map[string]*schema.Schema{
+		"format": {
+			Type:     schema.TypeString,
+			Optional: true,
+			Default:  "json",
+		},
+		"max_age": {
+			Type:     schema.TypeInt,
+			Optional: true,
+			Default:  30,
+		},
+		"max_backup": {
+			Type:     schema.TypeInt,
+			Optional: true,
+			Default:  10,
+		},
+		"max_size": {
+			Type:     schema.TypeInt,
+			Optional: true,
+			Default:  100,
+		},
+		"path": {
+			Type:     schema.TypeString,
+			Optional: true,
+			Default:  "/var/log/kube-audit/audit-log.json",
+		},
+		"policy": {
+			Type:     schema.TypeMap,
+			Optional: true,
+		},
+	}
+	return s
+}
+
+func clusterRKEConfigServicesKubeAPIAuditLogFields() map[string]*schema.Schema {
+	s := map[string]*schema.Schema{
+		"configuration": {
+			Type:     schema.TypeList,
+			MaxItems: 1,
+			Optional: true,
+			Elem: &schema.Resource{
+				Schema: clusterRKEConfigServicesKubeAPIAuditLogConfigFields(),
+			},
+		},
+		"enabled": {
+			Type:     schema.TypeBool,
+			Optional: true,
+			Default:  false,
+		},
+	}
+	return s
+}
+
+func clusterRKEConfigServicesKubeAPIEventRateLimitFields() map[string]*schema.Schema {
+	s := map[string]*schema.Schema{
+		"configuration": {
+			Type:     schema.TypeMap,
+			Optional: true,
+		},
+		"enabled": {
+			Type:     schema.TypeBool,
+			Optional: true,
+			Default:  false,
+		},
+	}
+	return s
+}
+
+func clusterRKEConfigServicesKubeAPISecretsEncryptionConfigFields() map[string]*schema.Schema {
+	s := map[string]*schema.Schema{
+		"custom_config": {
+			Type:     schema.TypeMap,
+			Optional: true,
+		},
+		"enabled": {
+			Type:     schema.TypeBool,
+			Optional: true,
+			Default:  false,
+		},
+	}
+	return s
+}
+
+func clusterRKEConfigServicesKubeAPIFields() map[string]*schema.Schema {
+	s := map[string]*schema.Schema{
+		"admission_configuration": {
+			Type:     schema.TypeMap,
+			Optional: true,
+		},
+		"always_pull_images": {
+			Type:     schema.TypeBool,
+			Optional: true,
+			Default:  false,
+		},
+		"audit_log": {
+			Type:     schema.TypeList,
+			MaxItems: 1,
+			Optional: true,
+			Elem: &schema.Resource{
+				Schema: clusterRKEConfigServicesKubeAPIAuditLogFields(),
+			},
+		},
+		"event_rate_limit": {
+			Type:     schema.TypeList,
+			MaxItems: 1,
+			Optional: true,
+			Elem: &schema.Resource{
+				Schema: clusterRKEConfigServicesKubeAPIEventRateLimitFields(),
+			},
+		},
+		"extra_args": {
+			Type:     schema.TypeMap,
+			Optional: true,
+			Computed: true,
+		},
+		"extra_binds": {
+			Type:     schema.TypeList,
+			Optional: true,
+			Elem: &schema.Schema{
+				Type: schema.TypeString,
+			},
+		},
+		"extra_env": {
+			Type:     schema.TypeList,
+			Optional: true,
+			Elem: &schema.Schema{
+				Type: schema.TypeString,
+			},
+		},
+		"image": {
+			Type:     schema.TypeString,
+			Optional: true,
+			Computed: true,
+		},
+		"pod_security_policy": {
+			Type:     schema.TypeBool,
+			Optional: true,
+			Default:  false,
+		},
+		"secrets_encryption_config": {
+			Type:     schema.TypeList,
+			MaxItems: 1,
+			Optional: true,
+			Elem: &schema.Resource{
+				Schema: clusterRKEConfigServicesKubeAPISecretsEncryptionConfigFields(),
+			},
+		},
+		"service_cluster_ip_range": {
+			Type:     schema.TypeString,
+			Optional: true,
+			Computed: true,
+		},
+		"service_node_port_range": {
+			Type:     schema.TypeString,
+			Optional: true,
+			Computed: true,
+		},
+	}
+	return s
+}
