+++ rancher2/schema_cluster_rke_config_services_kubelet.go
@@ -1,0 +1,62 @@
+package rancher2
+
+import (
+	"github.com/hashicorp/terraform-plugin-sdk/helper/schema"
+)
+
+//Schemas
+
+func clusterRKEConfigServicesKubeletFields() map[string]*schema.Schema {
+	s := map[string]*schema.Schema{
+		"cluster_dns_server": {
+			Type:     schema.TypeString,
+			Optional: true,
+			Computed: true,
+		},
+		"cluster_domain": {
+			Type:     schema.TypeString,
+			Optional: true,
+			Computed: true,
+		},
+		"extra_args": {
+			Type:     schema.TypeMap,
+			Optional: true,
+			Computed: true,
+		},
+		"extra_binds": {
+			Type:     schema.TypeList,
+			Optional: true,
+			Elem: &schema.Schema{
+				Type: schema.TypeString,
+			},
+		},
+		"extra_env": {
+			Type:     schema.TypeList,
+			Optional: true,
+			Elem: &schema.Schema{
+				Type: schema.TypeString,
+			},
+		},
+		"fail_swap_on": {
+			Type:     schema.TypeBool,
+			Optional: true,
+			Computed: true,
+		},
+		"generate_serving_certificate": {
+			Type:     schema.TypeBool,
+			Optional: true,
+			Default:  false,
+		},
+		"image": {
+			Type:     schema.TypeString,
+			Optional: true,
+			Computed: true,
+		},
+		"infra_container_image": {
+			Type:     schema.TypeString,
+			Optional: true,
+			Computed: true,
+		},
+	}
+	return s
+}
