+++ rancher2/structure_cluster_rke_config_services_kube_controller.go
@@ -1,0 +1,76 @@
+package rancher2
+
+import (
+	managementClient "github.com/rancher/types/client/management/v3"
+)
+
+// Flatteners
+
+func flattenClusterRKEConfigServicesKubeController(in *managementClient.KubeControllerService) ([]interface{}, error) {
+	obj := make(map[string]interface{})
+	if in == nil {
+		return []interface{}{}, nil
+	}
+
+	if len(in.ClusterCIDR) > 0 {
+		obj["cluster_cidr"] = in.ClusterCIDR
+	}
+
+	if len(in.ExtraArgs) > 0 {
+		obj["extra_args"] = toMapInterface(in.ExtraArgs)
+	}
+
+	if len(in.ExtraBinds) > 0 {
+		obj["extra_binds"] = toArrayInterface(in.ExtraBinds)
+	}
+
+	if len(in.ExtraEnv) > 0 {
+		obj["extra_env"] = toArrayInterface(in.ExtraEnv)
+	}
+
+	if len(in.Image) > 0 {
+		obj["image"] = in.Image
+	}
+
+	if len(in.ServiceClusterIPRange) > 0 {
+		obj["service_cluster_ip_range"] = in.ServiceClusterIPRange
+	}
+
+	return []interface{}{obj}, nil
+}
+
+// Expanders
+
+func expandClusterRKEConfigServicesKubeController(p []interface{}) (*managementClient.KubeControllerService, error) {
+	obj := &managementClient.KubeControllerService{}
+	if len(p) == 0 || p[0] == nil {
+		return obj, nil
+	}
+	in := p[0].(map[string]interface{})
+
+	if v, ok := in["cluster_cidr"].(string); ok && len(v) > 0 {
+		obj.ClusterCIDR = v
+	}
+
+	if v, ok := in["extra_args"].(map[string]interface{}); ok && len(v) > 0 {
+		obj.ExtraArgs = toMapString(v)
+	}
+
+	if v, ok := in["extra_binds"].([]interface{}); ok && len(v) > 0 {
+		obj.ExtraBinds = toArrayString(v)
+	}
+
+	if v, ok := in["extra_env"].([]interface{}); ok && len(v) > 0 {
+		obj.ExtraEnv = toArrayString(v)
+	}
+
+	if v, ok := in["image"].(string); ok && len(v) > 0 {
+		obj.Image = v
+	}
+
+	if v, ok := in["service_cluster_ip_range"].(string); ok && len(v) > 0 {
+		obj.ServiceClusterIPRange = v
+	}
+
+	return obj, nil
+}
