+++ rancher2/structure_cluster_rke_config_services_scheduler.go
@@ -1,0 +1,60 @@
+package rancher2
+
+import (
+	managementClient "github.com/rancher/types/client/management/v3"
+)
+
+// Flatteners
+
+func flattenClusterRKEConfigServicesScheduler(in *managementClient.SchedulerService) ([]interface{}, error) {
+	obj := make(map[string]interface{})
+	if in == nil {
+		return []interface{}{}, nil
+	}
+
+	if len(in.ExtraArgs) > 0 {
+		obj["extra_args"] = toMapInterface(in.ExtraArgs)
+	}
+
+	if len(in.ExtraBinds) > 0 {
+		obj["extra_binds"] = toArrayInterface(in.ExtraBinds)
+	}
+
+	if len(in.ExtraEnv) > 0 {
+		obj["extra_env"] = toArrayInterface(in.ExtraEnv)
+	}
+
+	if len(in.Image) > 0 {
+		obj["image"] = in.Image
+	}
+
+	return []interface{}{obj}, nil
+}
+
+// Expanders
+
+func expandClusterRKEConfigServicesScheduler(p []interface{}) (*managementClient.SchedulerService, error) {
+	obj := &managementClient.SchedulerService{}
+	if len(p) == 0 || p[0] == nil {
+		return obj, nil
+	}
+	in := p[0].(map[string]interface{})
+
+	if v, ok := in["extra_args"].(map[string]interface{}); ok && len(v) > 0 {
+		obj.ExtraArgs = toMapString(v)
+	}
+
+	if v, ok := in["extra_binds"].([]interface{}); ok && len(v) > 0 {
+		obj.ExtraBinds = toArrayString(v)
+	}
+
+	if v, ok := in["extra_env"].([]interface{}); ok && len(v) > 0 {
+		obj.ExtraEnv = toArrayString(v)
+	}
+
+	if v, ok := in["image"].(string); ok && len(v) > 0 {
+		obj.Image = v
+	}
+
+	return obj, nil
+}
